requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-JWT-Extended==3.12.0
Flask-RESTful==0.3.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
py-bcrypt==0.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
